Update Home.vue
Fixed check on scrollPosition when mounted, this might indicate that the current user has uploaded new content. In order to avoid the name(s) on the new content not being shown if the user hasn't been active before get all the names. Because the getData function might not be called depending on where the lazy loader trigger is in the viewport / all the content has already been loaded.